+++ .gitignore
@@ -1,79 +1,79 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
+ 
 # dependencies
 /node_modules
 /.pnp
 .pnp.js
 .yarn/install-state.gz
-
+ 
 # testing
 /coverage
-
+ 
 # next.js
 /.next/
 /out/
-
+ 
 # production
 /build
-
+ 
 # misc
 .DS_Store
 *.pem
-
+ 
 # debug
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
+ 
 # local env files
 .env*.local
-
+ 
 # vercel
 .vercel
-
+ 
 # typescript
 *.tsbuildinfo
 next-env.d.ts
-
+ 
 .next
 node_modules
-
+ 
 .env
-
+ 
 # Created by https://www.gitignore.io
-
+ 
 ### OSX ###
 .DS_Store
 .AppleDouble
 .LSOverride
-
+ 
 # Icon must end with two \r
 Icon
-
-
+ 
+ 
 # Thumbnails
 ._*
-
+ 
 # Files that might appear on external disk
 .Spotlight-V100
 .Trashes
-
+ 
 # Directories potentially created on remote AFP share
 .AppleDB
 .AppleDesktop
 Network Trash Folder
 Temporary Items
 .apdisk
-
-
+ 
+ 
 ### Python ###
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
-
+ 
 # C extensions
 *.so
-
+ 
 # Distribution / packaging
 .Python
 env/
@@ -90,17 +90,17 @@ var/
 *.egg-info/
 .installed.cfg
 *.egg
-
+ 
 # PyInstaller
 #  Usually these files are written by a python script from a template
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
 *.manifest
 *.spec
-
+ 
 # Installer logs
 pip-log.txt
 pip-delete-this-directory.txt
-
+ 
 # Unit test / coverage reports
 htmlcov/
 .tox/
@@ -108,24 +108,24 @@ htmlcov/
 .cache
 nosetests.xml
 coverage.xml
-
+ 
 # Translations
 *.mo
 *.pot
-
+ 
 # Sphinx documentation
 docs/_build/
-
+ 
 # PyBuilder
 target/
-
-
+ 
+ 
 ### Django ###
 *.log
 *.pot
 *.pyc
 __pycache__/
 local_settings.py
-
+ 
 .env
 db.sqlite3
