+++ Backend/backend/db_connection.py
@@ -7,8 +7,7 @@ load_dotenv()
 DATABASE_USER = os.getenv("MONGO_USER")
 DATABASE_PASSWORD = os.getenv("MONGO_PASSWORD")
 
-url = f'mongodb+srv://{DATABASE_USER}:{
-    DATABASE_PASSWORD}@app-cluster.kty585r.mongodb.net/?retryWrites=true&w=majority'
+url = f"mongodb+srv://{DATABASE_USER}:{DATABASE_PASSWORD}@app-cluster.kty585r.mongodb.net/?retryWrites=true&w=majority"
 
 client = pymongo.MongoClient(url)
