+++ Frontend/app/globals.css
@@ -49,29 +49,41 @@ body {
   width: 100%;
   height: 100%;
   margin: 0 auto;
+
 }
 
 .dashboard {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
-  justify-content: center;
+  justify-content: space-between;
+  justify-items: center;
   align-items: center;
-  gap: 150px;
-  border: 1px solid black;
-  padding: 16px;
+  align-content: stretch;
+  gap: 100px;
+  /* border: 1px solid black; */
+  padding: 20px;
   margin: 0;/* Use flex: 1 to make the dashboard fill the remaining height of the layout */
-  margin-top: 150px;
-  border-radius: 20px;
+  margin-top: 0px;
+  border-radius: 100px;
   background-color: #fff;
+  margin-top: 50px;
+  place-content: space-between;
+  height: 90vh;
+  width: fit-content;
+  height: fit-content;
+  box-shadow: 0 1px 3px 0 rgb(0 0 0 / 0.1), 0 1px 2px -1px rgb(0 0 0 / 0.1);
 }
 
 .boxes {
-  color: black;
-  border: 1px solid black; /* Add border color for the boxes */
+  background: white;
+  --tw-shadow-color: #431407;
+  /* border: 1px solid black; */
   padding: 150px; /* Add padding for spacing inside the boxes */
   text-align: center; /* Center text inside the boxes */
   border-radius: 50px;
-  background-color: blue;
+  height: 40vh;
+  width: 700px;
+  box-shadow: 0 25px 50px -12px rgb(0 0 0 / 0.25);
 }
 
 @layer utilities {
@@ -185,4 +197,9 @@ sidebar_name {
 .button:focus-visible {
   outline: 2px solid #4f46e5; /* Equivalent to 'focus-visible:outline' */
   outline-offset: 2px; /* Equivalent to 'focus-visible:outline-offset-2' */
+}
+
+.bar {
+  width: "100px";
+  height: "4000px";
 }
