+++ Frontend/app/register/page.tsx
@@ -1,9 +1,40 @@
 "use client";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
+import { useState } from "react";
+import database from "../api/database";
+import axios from "axios";
 
 export default function Signup() {
+  const [firstName, setFirstName] = useState('');
+  const [lastName, setLastName] = useState('');
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
   const router = useRouter();
+
+
+  const handleSubmit = async () => {
+    database.post('/register', {
+      firstName: firstName,
+      lastName: lastName,
+      emailAddress: email,
+      password: password
+    })
+    .then(function (response) {
+      console.log(response);
+    })
+    .catch(function (error) {
+      console.log(error);
+    });
+
+    database.post('/login', {
+      emailAddress: email,
+      password: password
+    })
+
+    router.push('/dashboard')
+  }
 
   return (
     <div className="flex justify-center mt-8">
@@ -35,6 +66,7 @@ tracking-tight text-gray-900">
                     name="firstname"
                     type="firstname"
                     autoComplete="firstname"
+                    onChange={(event) => setFirstName(event.target.value)}
                     required
                     className="block w-full rounded-md border-0 py-1.5 
 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 
@@ -56,6 +88,7 @@ sm:text-sm sm:leading-6"
                     name="lastname"
                     type="lastname"
                     autoComplete="lastname"
+                    onChange={(event) => setLastName(event.target.value)}
                     required
                     className="block w-full rounded-md border-0 py-1.5 
 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 
@@ -77,6 +110,7 @@ sm:text-sm sm:leading-6"
                     name="email"
                     type="email"
                     autoComplete="email"
+                    onChange={(event) => setEmail(event.target.value)}
                     required
                     className="block w-full rounded-md border-0 py-1.5 
 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 
@@ -101,6 +135,7 @@ sm:text-sm sm:leading-6"
                     name="password"
                     type="password"
                     autoComplete="current-password"
+                    onChange={(event) => setPassword(event.target.value)}
                     required
                     className="block w-full rounded-md border-0 py-1.5 text-gray-900 
 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 
@@ -124,6 +159,7 @@ focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
                     name="confirmpassword"
                     type="password"
                     autoComplete="current-password"
+                    onChange={(event) => setConfirmPassword(event.target.value)}
                     required
                     className="block w-full rounded-md border-0 py-1.5 
 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 
@@ -137,7 +173,7 @@ sm:text-sm sm:leading-6"
                 <button
                   type="submit"
                   className="button"
-                  onClick={() => router.push('/')}
+                  onClick={handleSubmit}
                 >
                   Sign in
                 </button>
@@ -147,9 +183,7 @@ sm:text-sm sm:leading-6"
             <p className="mt-10 text-center text-sm text-gray-500">
               Already have an account?{" "}
               <span
-                onClick={() => {
-                  router.push("/");
-                }}
+                onClick={() => router.push('/')}
                 className="font-semibold leading-6 text-indigo-600 
 hover:text-indigo-500 cursor-pointer"
               >
